LayoutTests/svg/dynamic-updates/SVGFETileElement-svgdom-in-prop-expected.txt

CONSOLE MESSAGE: line 62: TypeError: 'undefined' is not an object (evaluating 'tile.in1.baseVal = "offset"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFETileElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL tile.in1.baseVal should be SourceGraphic. Threw exception TypeError: 'undefined' is not an object (evaluating 'tile.in1.baseVal')
